refactor: rename components in PascalCase
