+++ src/components/home/ProjectGroupCard.tsx
