+++ src/components/home/SectionSocials.tsx
