+++ src/components/home/SlideInfo.tsx
