+++ src/components/home/SlideMission.tsx
